Update screen size and video duration types

+++ dashboard/backend/Application/VideoSessions/Commands/CreateVideoSession/CreateVideoSessionCommand.cs
@@ -8,6 +8,6 @@ namespace Application.VideoSessions.Commands.CreateVideoSession
         public required Guid SessionID { get; set; }
         public string? VideoID { get; set; }
         public required string Source { get; set; }
-        public required int Duration { get; set; }
+        public required double Duration { get; set; }
     }
 }
