+++ dashboard/backend/Application/VideoSessions/Commands/EndVideoSession/EndVideoSessionCommand.cs
@@ -8,6 +8,6 @@ namespace Application.VideoSessions.Commands.EndVideoSession
         public required Guid WebsiteKey { get; set; }
         public required Guid VideoSessionID { get; set; }
         public required Guid SessionID { get; set; }
-        public required int Duration { get; set; }
+        public required double Duration { get; set; }
     }
 }
