+++ dashboard/backend/Domain/Entities/VideoSession.cs
@@ -8,7 +8,7 @@ namespace Domain.Entities
         public required Guid SessionId { get; set; }
         public string VideoId { get; set; } = "Not specified";
         public required string Source { get; set; }
-        public required int Duration { get; set; }
+        public required double Duration { get; set; }
         public Session? Session { get; set; }
         public ICollection<VideoEvent>? VideoEvents { get; set; }
     }
